requirements.txt
python-telegram-bot==20.7
yfinance
requests
numpy
apscheduler
reportlab
